Move auth.dto.UserRegistrationDto -> user.dto.UserRegistrationDto, auth.dto.UpdateDto -> user.dto.UpdateDto
