+++ src/main/java/site/hirecruit/hr/domain/auth/service/UserUpdateService.kt
@@ -1,7 +1,7 @@
 package site.hirecruit.hr.domain.auth.service
 
 import site.hirecruit.hr.domain.auth.dto.AuthUserInfo
-import site.hirecruit.hr.domain.auth.dto.UserUpdateDto
+import site.hirecruit.hr.domain.user.dto.UserUpdateDto
 
 /**
  * User의 기본정보를 Update하는 Service
