+++ src/main/java/site/hirecruit/hr/domain/user/dto/UserRegistrationDto.kt
@@ -1,4 +1,4 @@
-package site.hirecruit.hr.domain.auth.dto
+package site.hirecruit.hr.domain.user.dto
 
 import com.fasterxml.jackson.annotation.JsonGetter
 import com.fasterxml.jackson.annotation.JsonProperty
